Extracted some fixtures
