+++ tests/integration/test_ssh.py
@@ -1,22 +1,12 @@
 import asyncio
 import io
-from pathlib import PurePath
 from textwrap import dedent
 from typing import NamedTuple
 
 import aiodocker.utils
 import asyncssh
 import pytest
-from yarl import URL
 
-from platform_api.config import (
-    Config,
-    DatabaseConfig,
-    IngressConfig,
-    LoggingConfig,
-    ServerConfig,
-    StorageConfig,
-)
 from platform_api.orchestrator.job import JobRequest
 from platform_api.orchestrator.job_request import Container, ContainerResources
 from platform_api.orchestrator.kube_orchestrator import KubeOrchestrator, PodDescriptor
@@ -45,28 +35,6 @@ class ApiConfig(NamedTuple):
     @property
     def ping_url(self):
         return self.endpoint + "/ping"
-
-
-@pytest.fixture
-def config(kube_config, redis_config, auth_config, es_config):
-    server_config = ServerConfig()
-    storage_config = StorageConfig(host_mount_path=PurePath("/tmp"))  # type: ignore
-    database_config = DatabaseConfig(redis=redis_config)  # type: ignore
-    logging_config = LoggingConfig(elasticsearch=es_config)
-    ingress_config = IngressConfig(
-        storage_url=URL("https://neu.ro/api/v1/storage"),
-        users_url=URL("https://neu.ro/api/v1/users"),
-        monitoring_url=URL("https://neu.ro/api/v1/monitoring"),
-    )
-    return Config(
-        server=server_config,
-        storage=storage_config,
-        orchestrator=kube_config,
-        database=database_config,
-        auth=auth_config,
-        logging=logging_config,
-        ingress=ingress_config,
-    )
 
 
 @pytest.fixture
